+++ src/components/DatePicker/SingleDatePicker.tsx
@@ -1,19 +1,19 @@
 'use client'
 import { FC, createElement, forwardRef, useState } from 'react'
-import DatePicker from 'react-datepicker'
+import SingleDate from 'react-datepicker'
 import { useDatePickerContext } from './DatePickerContext'
 import { CustomInput } from './CustomInput'
 
 export const SingleDatePicker: FC = () => {
-  const { showTwoMonth, singleDatePicker, placeholder } = useDatePickerContext()
+  const { showTwoMonth, singleDate, placeholder } = useDatePickerContext()
   const [date, setDate] = useState<Date | null>()
   const handleChange = (date: Date | null) => {
     setDate(date)
-    singleDatePicker && singleDatePicker(date)
+    singleDate && singleDate(date)
   }
 
   return (
-    <DatePicker
+    <SingleDate
       placeholderText={placeholder}
       monthsShown={showTwoMonth ? 2 : 1}
       selected={date}
@@ -21,7 +21,7 @@ export const SingleDatePicker: FC = () => {
       customInput={createElement(forwardRef(CustomInput))}
       className="rounded-md border border-metal-300 placeholder:text-body-5"
       showPopperArrow={false}
-      startDate={date}
-      dateFormat="dd/MM/yyyy"></DatePicker>
+      dateFormat="dd/MM/yyyy"
+    />
   )
 }