+++ src/components/DatePicker/TimePicker.tsx
@@ -1,11 +1,11 @@
 'use client'
 import { FC, createElement, forwardRef, useState } from 'react'
-import DatePicker from 'react-datepicker'
+import TimePick from 'react-datepicker'
 import { useDatePickerContext } from './DatePickerContext'
 import { CustomInput } from './CustomInput'
 
 export const TimePicker: FC = () => {
-  const { timePicker, placeholder } = useDatePickerContext()
+  const { time, placeholder } = useDatePickerContext()
   const [selectedTime, setSelectedTime] = useState(new Date())
 
   const handleTimePicker = (date: Date) => {
@@ -21,11 +21,11 @@ export const TimePicker: FC = () => {
     const amOrPm = hours < 12 ? 'am' : 'pm'
     const formattedMinutes = String(minutes).padStart(2, '0')
     const newTime = `${formattedHours}:${formattedMinutes}${amOrPm}`
-    timePicker && timePicker(newTime)
+    time && time(newTime)
   }
 
   return (
-    <DatePicker
+    <TimePick
       placeholderText={placeholder}
       customInput={createElement(forwardRef(CustomInput))}
       selected={selectedTime}