+++ src/components/DatePicker/YearPicker.tsx
@@ -1,23 +1,23 @@
 'use client'
 import { CaretLeft, CaretRight } from 'phosphor-react'
 import { createElement, forwardRef, useState } from 'react'
-import ReactDatePicker from 'react-datepicker'
+import YearPick from 'react-datepicker'
 import { useDatePickerContext } from './DatePickerContext'
 import { CustomInput } from './CustomInput'
 
 export const YearPicker: React.FC = () => {
-  const { yearPicker, placeholder } = useDatePickerContext()
+  const { year, placeholder } = useDatePickerContext()
   const [startDate, setStartDate] = useState<Date | null>(null)
   const handleDateChange = (date: Date | null) => {
     if (date) {
       setStartDate(date)
-      const year = new Date(date).getFullYear()
-      yearPicker && yearPicker(year)
+      const newYear = new Date(date).getFullYear()
+      year && year(newYear)
     }
   }
 
   return (
-    <ReactDatePicker
+    <YearPick
       placeholderText={placeholder}
       selected={startDate}
       onChange={handleDateChange}